contact adding feature

+++ src/components/EditableDataComponent.js
@@ -7,18 +7,29 @@ import { ModalDispatchContext } from "./context_providers/DataProvider";
 import Form from "./Form";
 import DeletionConfirmation from "./DeletionConfirmation";
 
-export default function EditableDataComponent({ inputsData, className, onUpdate, onDeletion, children }) {
+export default function EditableDataComponent({ inputsData, className, onUpdate, onDeletion, onAddition, children }) {
 	const [isHovered, setHover] = useState(false);
 	const dispatchModal = useContext(ModalDispatchContext);
 	const editForm = <Form inputsDataArr={inputsData} onSubmit={handleSubmit} />;
-	const deletionConfirmation = <DeletionConfirmation onAccept={handleDeletion} onCancel={handleModalClosing} />;
 	const editButton = createButton({ type: "edit", onClick: () => openModal(editForm) });
-	const deleteButton = createButton({ type: "delete", onClick: () => openModal(deletionConfirmation) });
+
+	let deletionConfirmation = null;
+	let deleteButton = null;
+	if (onDeletion) {
+		deletionConfirmation = <DeletionConfirmation onAccept={handleDeletion} onCancel={handleModalClosing} />;
+		deleteButton = createButton({ type: "delete", onClick: () => openModal(deletionConfirmation) });
+	}
+
+	let addButton = null;
+	if (onAddition) {
+		addButton = createButton({ type: "add", onClick: onAddition });
+	}
 
 	return (
 		<EditableDataComponentStyled className={className} onMouseOver={handleMouseOver} onMouseOut={handleMouseOut}>
 			{children}
 			<EditPanelLayout className={(isHovered) ? "active" : "inactive"}>
+				{addButton}
 				{editButton}
 				{deleteButton}
 			</EditPanelLayout>
@@ -64,8 +75,12 @@ export default function EditableDataComponent({ inputsData, className, onUpdate,
 
 EditableDataComponent.propTypes = {
 	inputsData: propTypes.array,
-	children: propTypes.object,
+	children: propTypes.oneOfType([
+		propTypes.object,
+		propTypes.array
+	]),
 	className: propTypes.string,
 	onUpdate: propTypes.func,
-	onDeletion: propTypes.func
+	onDeletion: propTypes.func,
+	onAddition: propTypes.func
 };
